Budget blocker Error removed

+++ NPD/EmcureNPD.Web/Controllers/ManagementController.cs
@@ -266,9 +266,9 @@ namespace EmcureNPD.Web.Controllers
                         model.MarketingAllowance = Convert.ToDecimal(data.table[0].marketingAllowance);
                         model.RegulatoryMaintenanceCost = Convert.ToDecimal(data.table[0].regulatoryMaintenanceCost);
                         model.GrosstoNet = Convert.ToDecimal(data.table[0].grosstoNet);
-                        model.Noofbatchestobemanufactured = Convert.ToInt32(data.table[0].noofbatchestobemanufactured);
+                        model.Noofbatchestobemanufactured = Convert.ToDouble(data.table[0].noofbatchestobemanufactured);
                         model.NoofbatchestobemanufacturedPhaseEndDate = data.table[0].noofbatchestobemanufacturedPhaseEndDate;
-                        model.NoSkus = Convert.ToInt32(data.table[0].noSkus);
+                        model.NoSkus = Convert.ToDouble (data.table[0].noSkus);
                         model.NoSkusPhaseEndDate = data.table[0].noSkusPhaseEndDate;
                         model.RandDanalyticalcost = Convert.ToDecimal(data.table[0].randDanalyticalcost);
                         model.RandDanalyticalcostPhaseEndDate = data.table[0].randDanalyticalcostPhaseEndDate;
